DYN-7899: Type forward nodecategoryattribute for use in external node libs such as protogeometry (#16914)
