+++ src/DynamoCore/Properties/AssemblyInfo.cs
@@ -64,6 +64,8 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("MCPServer")]
 
 // Disable PublicAPIAnalyzer errors for this type as they're already added to the public API text file
-#pragma warning disable RS0016 
+#pragma warning disable RS0016
 [assembly: TypeForwardedTo(typeof(Dynamo.Scheduler.Disposable))]
+// NodeCategoryAttribute moved to DynamoServices so libraries that reference only DynamoServices (e.g. LibG) can use it.
+[assembly: TypeForwardedTo(typeof(Dynamo.Graph.Nodes.NodeCategoryAttribute))]
 #pragma warning restore RS0016
