+++ src/NodeServices/DynamoServices.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
     <ImportGroup Label="PropertySheets">
         <Import Project="$(SolutionDir)Config\CS_SDK.props" />
     </ImportGroup>
@@ -19,9 +19,13 @@
         <TargetFrameworks>netstandard2.0;net8.0;net10.0</TargetFrameworks>
     </PropertyGroup>
     <PropertyGroup>
-        <NoWarn>MSB3539;CS1591;NUnit2005;NUnit2007;CS0618;CS0612;CS0672</NoWarn>
+        <NoWarn>MSB3539;CS1591;NUnit2005;NUnit2007;CS0618;CS0612;CS0672;RS0026</NoWarn>
     </PropertyGroup>
     <ItemGroup>
+        <PackageReference Include="Microsoft.CodeAnalysis.PublicApiAnalyzers" Version="3.3.4">
+            <PrivateAssets>all</PrivateAssets>
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+        </PackageReference>
         <!--exclude these dependencies as Sys.Ref.MetadataLoadContext.dll is already pulled in from other projects and the transitive deps are part of the .net8/net10 runtime.-->
         <PackageReference Include="System.Reflection.MetadataLoadContext" Version="8.0.0" ExcludeAssets="runtime" />
     </ItemGroup>
